refactor: 포켓볼, 버튼 컴포넌트 세부 분리
